Create BoxStacker.c
A recursive function designed to stack boxes without rotation.  To be considered for the stack, a box must be less in all dimensions to the current stack top (unless there is no box in the stack).

This solution uses a dynamic approach with a sort function by maximum height of taking and leaving the box, for each valid box.